requirements.txt
# MongoDB
pymongo==4.5.0
motor==3.3.1  # 비동기 MongoDB 드라이버 (선택적)

# Oracle
cx-Oracle==8.3.0
SQLAlchemy==2.0.23

# 환경 설정
python-dotenv==1.0.0

# 로깅
loguru==0.7.2  # 선택적: 더 나은 로깅

# 데이터 처리
pandas==2.1.3  # 선택적: 복잡한 데이터 처리
numpy==1.24.3  # 선택적: 수치 처리

# 비동기 처리
asyncio==3.4.3
aiofiles==23.2.1  # 선택적: 비동기 파일 처리

# 테스트
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
faker==20.0.3  # 테스트 데이터 생성

# 타입 체킹
mypy==1.7.0
types-pymongo==4.5.0

# 코드 품질
black==23.11.0
flake8==6.1.0
pylint==3.0.2

# Airflow (선택적)
apache-airflow==2.7.3